[Fix] 게시글 및 댓글 리턴 타입 수정

+++ bang2bwa/src/main/java/com/ssafy/bbb/model/dao/CommentDao.java
@@ -7,7 +7,7 @@ import com.ssafy.bbb.model.dto.notice.CommentResponseDto;
 
 public interface CommentDao {
 	// 댓글 작성
-	int insert(CommentRequestDto request);
+	Long insert(CommentRequestDto request);
 	
 	// 댓글 조회
 	List<CommentResponseDto> commentList(Long postId);
